+++ commerce/payment/src/main/java/ru/yandex/practicum/service/PaymentService.java
@@ -3,15 +3,16 @@ package ru.yandex.practicum.service;
 import ru.yandex.practicum.dto.OrderDto;
 import ru.yandex.practicum.dto.PaymentDto;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface PaymentService {
 
     PaymentDto createPayment(OrderDto order);
 
-    double calculateProductCost(OrderDto order);
+    BigDecimal calculateProductCost(OrderDto order);
 
-    double calculateTotalCost(OrderDto order);
+    BigDecimal calculateTotalCost(OrderDto order);
 
     void setPaymentSuccessful(UUID paymentId);
